Added all panels for rough draft sign up and sql database structure
I have completed the three panels for the rough draft for signing up for
the auction service. These gui elements are not "smart" and will need
additional code to check for duplicate entries in the database, correct
format of input, etc. I have also included a rough draft of the sql file
containing the structure of the database we plan to use